+++ CHANGELOG.md
@@ -73,6 +73,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 * Jaeger will get random free ports exposed for HOST in `docker-compose`
 * More flexible `MonetaryAmount` VO
 * Simplified mapping
+* Deafult permissions will be the same for both GETS and use the entity plural ((e.g. `CanReadRecipes` ). All permissions will use the entity plural.
 
 ### Fixed